Create cmd/root.go
